Seperate Active Abilities
Active abilities are now seperated be triggered by numbers 1-4. There is no UI yet, which Nick and I will need to implement on Friday.

***Known Issue***
I had to put BaseAbility and ActiveAbility under the Powerups namespace, which honestly should have been the case to begin with. This means they now spawn as pickups in their base form sometimes, which will need to be fixed.